前台後台 登入驗證 基本CRUD

+++ Booking/Models/Activity.cs
@@ -9,7 +9,7 @@ public partial class Activity
 
     public string Name { get; set; } = null!;
 
-    public byte Company { get; set; }
+    public short Company { get; set; }
 
     public string? PhotoUrl { get; set; }
